Move the base ClassLoader into the Garden namespace
